+++ config/routes.rb
@@ -1,4 +1,6 @@
 EasyTokens::Engine.routes.draw do
   root to: 'tokens#index'
-  resources :tokens, except: [:destroy]
+  resources :tokens, except: [:show, :destroy] do
+    get :deactivate_token
+  end
 end