+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
-# 0.3.0 - 2014-11-29
-## Added
+## 0.4.0 - 2015-12-04
+
+### Added
+
+* Change the views layout (custom CSS)
+* Remove redundant show view
+* Adds deactivate token action
+
+## 0.3.0 - 2014-11-29
+
+### Added
+
 * Initial working version
 * Mountable engine with basic CRUD for interaction with Token objects
 * `check_token_on` method available in controller